requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==0.11.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
stripe==7.8.0
redis==5.0.1
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
twisted==23.10.0
python-multipart==0.0.6
python-dotenv==1.0.0